Fix and clarify cumulative maximum/symmetry breaking comments and code
Clarify the comments describing the cumulative maximum array's purpose and the
conditional that determines whether to fix it up. Initialize the array with a
value that satisfies the constraint in the first place. Shorten up the setting
of the array's value for the current node. Set the array's value for the next
node before the recursive call for the next node. Don't propagate the
constraints if we're already at the last node or the last color.